Add WIP dynamic precache system
It has a custom binary format to store bumps, but currently there is no action implemented that causes a bump, so it will generate empty.

The contents of the bump file can be set to this hexdump:
01000000FB2E7564000000006461666679000000000000000000000000000000000000000000000000000000
This causes 1 bump to be present, daffy. The random-looking bytes are a timestamp, and the code will discard this bump a month from now.